refactor graphql call out
